fix: do even more adjustments
